added tooltip to the profile picture

+++ src/pages/myToys/MyToys.jsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../providers/AuthProvider";
-import { FaRegEdit, FaTimes } from "react-icons/fa";
+import MyToy from "./MyToy";
 
 
 const MyToys = () => {
@@ -10,27 +10,30 @@ const MyToys = () => {
     const { email } = user;
     console.log(email)
     useEffect(() => {
-        fetch(`http://localhost:5000/toys?email=${email}`)
+        fetch(`https://dream-motorz-server.vercel.app/toys?email=${email}`)
             .then(res => res.json())
             .then(data => {
                 setMyToys(data);
             })
     }, [email])
 
-    const handleDelete = (id) =>{
-        fetch(`http://localhost:5000/products/${id}`, {
-            method: 'DELETE'
-        })
-        .then(res => res.json())
-        .then(data => {
-            console.log(data);
-            if(data.deletedCount > 0){
-                const remaining = myToys.filter(toy => toy._id !== id);
-                setMyToys(remaining);
-            }
-        })
-    }
+    const handleDelete = (id) => {
+        const asking = confirm('Are you sure to delete?');
+        if (asking) {
+            fetch(`https://dream-motorz-server.vercel.app/products/${id}`, {
+                method: 'DELETE'
+            })
+                .then(res => res.json())
+                .then(data => {
+                    console.log(data);
+                    if (data.deletedCount > 0) {
+                        const remaining = myToys.filter(toy => toy._id !== id);
+                        setMyToys(remaining);
+                    }
+                })
+        }
 
+    }
 
     return (
         <div className="overflow-x-auto w-full md:p-4">
@@ -50,33 +53,13 @@ const MyToys = () => {
                 <tbody>
                     {/* row 1 */}
                     {
-                        myToys.map(toy => <tr key={toy._id}>
-                            <td>{toy.sellerName}</td>
-                            <td>{toy.toyName}</td>
-                            <td>{toy.subCategory}</td>
-                            <td>${toy.price}</td>
-                            <td>{toy.availableQuantity}</td>
-                            <td>
-                                <div className="flex space-x-2">
-                                    <button className="btn btn-circle bg-amber-500 hover:bg-amber-600 border-none text-xl"><FaRegEdit /></button>
-                                    <button onClick={()=> handleDelete(toy._id)} className="btn btn-circle bg-red-500 hover:bg-red-600 border-none text-xl"><FaTimes /></button>
-                                </div>
-                            </td>
-                        </tr>)
+                        myToys.map(toy => <MyToy
+                            key={toy._id}
+                            toy={toy}
+                            handleDelete={handleDelete}
+                        ></MyToy>)
                     }
                 </tbody>
-                {/* foot */}
-                <tfoot>
-                    <tr>
-                        <th>Seller</th>
-                        <th>Toy Name</th>
-                        <th>Sub-category</th>
-                        <th>Price</th>
-                        <th>Available Quantity</th>
-                        <th>Actions</th>
-                    </tr>
-                </tfoot>
-
             </table>
         </div>
     );
